SPORTS-73 Move inline styles to scss file
