--- frontend/src/components/Test.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function Player(props) {
-  return <div style={{ background: 'red' }}>This is players page</div>;
-}
