+++ frontend/src/components/hocs/redirectIfAuthenticated.js
@@ -3,10 +3,14 @@ import PropsTypes from 'prop-types';
 import { Redirect } from 'react-router';
 import { branch, renderComponent } from 'recompose';
 
+import { ADMIN_TOURNAMENTS } from '../../constants/routes';
+
 const checkAuthentication = ({ isAuthenticated }) => isAuthenticated;
 
 const RedirectHere = ({ location }) => {
-  const newRoute = (location.state && location.state.from) || { pathname: '/' };
+  const newRoute = (location.state && location.state.from) || {
+    pathname: ADMIN_TOURNAMENTS
+  };
 
   return <Redirect to={newRoute} />;
 };
