+++ frontend/src/components/players/index.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export default function Players(props) {
+  return <div>This is players page</div>;
+}
