+++ frontend/webpack.config.js
@@ -51,7 +51,7 @@ module.exports = {
         use: ['file-loader']
       },
       {
-        test: /\.(png|woff|woff2|eot|ttf|svg)$/,
+        test: /\.(woff|woff2|eot|ttf|svg)$/,
         loader: 'url-loader?limit=100000'
       }
     ]
